DOC: adding template to show Zenko forums in pages
This template adds the possibility to discuss documentation on
Zenko forum. At the bottom of each docs page an iframe is loaded
showing a prompt to start a new thread in case there is not one
already. If there is already a thread open, the associated
discussion will be loaded.