rf: ZENKO-758 rename replication pods
Add "replication" to pod names to avoid confusions, and change
"producer" and "consumer" to more meaningful names.

New pod names:

- backbeat-producer -> backbeat-replication-populator
- backbeat-consumer -> backbeat-replication-data-processor
- backbeat-status -> backbeat-replication-status-processor

If there's a risk it does not fit anymore in DNS name limit (63 char?)
we can consider shortening names, but ideally (to me) they would be as
self-descriptive as they can get.