rf: remove wait flags and programatically wait
Reduces flakiness with the installation step of the CI where sometime pods would be in indefinitate Crashloop
state that would cause the CI potentially wait 800 seconds for an install to finish. The Crashloop state is normal
especially in the CI environment where the services are expecting the buckets that are predefined environment but dont
yet exist until test initialization. The CI still explicitly expects the stack to stabilize but this is now part of
the test initialization thus saving time on sleep commands and unneccessary waits. Additionally, this allows devs to
collect logs on a stack that was unable to stabilize for debugging purposes.