ZENKO-1085 bugfix: backbeat liveness probes
There is no point on having readiness probes for
the Backbeat pods since they don't serve requests
(except for BackbeatAPI).

Also, the readiness route is more appropriate for
the liveness probe since it detects when a connection
to a dependency (i.e. Zookeeper) is lost and
Kubernetes can take advantage of that to restart
pods and allow them to reconnect properly. This is
necessary since the Backbeat pods don't seem to recover
on their own from these disconnections.