ci: fix false positives on stability checks
This commit fixes an issue where a CI build was allowed to proceed with
tests before being fully stabilized. This is due to the short window where a
container restarts and is momentarily ready before crashing for whatever
reason. The fix is to add a stabilization period of 10 second where everything
is expected to be ready.