bugfix: ZENKO-1826 set UV_THREADPOOL_SIZE in data processor pod
Set the UV_THREADPOOL_SIZE environment variable to 64 (default is 4)
in the data processor pods, to fix replication tasks being very slow
when the number of zenko locations is large (~20).

The current theory explaining the fix: the libuv thread pool is
normally used when doing external processing with significant CPU use,
so not to block the main event loop. 4 threads is apparently not
enough to handle the load generated by KafkaConsumer instances (using
librdkafka underneath), and it slows down the processing of
http.request() calls. I assume things must get queued faster than they
can be processed, resulting in very large delays.

Choosing 64 as a heuristic, 128 is the hard maximum we can set. We can
experiment with this value and change it later if it causes issues.