ci: wait for all stateful services
This commit staggers the stabilization by waiting in appropriate order prior to
attempting bucket creation. Also this waits for all the stateful services to
become available prior to waiting for pod stabilization. While this can cause
higher setup time the benefit allows for a far more reliable and consistent
installation setup and stablization.