documentation: ZENKOIO-96 Fix API Headings (3rd try)
This PR does a whole lot of reorganization and orthographic fixes on the
Reference guide.

- Requests and responses given heading status.
- API calls & responses rendered in monospace.
- Tables fixed to be legible in both HTML and PDF outputs.
- Backbeat and Prometheus rendered more like the other APIs in the book.

This fixes #ZENKOIO-96

Note: This is a big one. Sorry for the volume to review. I've discussed this
with the DE, and further work is required on ordering Backbeat and Prometheus
calls into a user-friendly format (i.e., as being accessible through Cloudserver
API, rather than as freestanding APIs). I'll create another ticket for this set
of edits.