bf: add s3 version IDs on mdonly
This commit fixes some issues in the RClone patches that produce a map of the
metadata headers of an S3 request. Originally and incorrectly, the map of the
metadata should have been in a map[string]*string data structure (array of
string pointers). In fixing this, I was able to add the versionID in the
readMetadata method to the *Object.meta parameter and pass it along with the
request to Cloudserver.