charts: Update apiVersions to support K8s 1.16
Many deprecated APIs were removed in Kubernetes 1.16 (see:
https://kubernetes.io/blog/2019/07/18/api-deprecations-in-1-16/)

Since all the charts in this repository depend on K8s 1.10 or above, we
migrate all resources in the affected APIs to the latest apiVersion
available in K8s 1.10. This entirely covers the needs to support K8s
1.16.

For Deployments/StatefulSets that were missing a `selector`, we add one
(since it is now required, and immutable, in the apps/v1 API).

Fixes: ZENKO-2574