improvement/ZENKO-2992_remove_orbit
Removed final Orbit UI elements. In the process, discovered that operations
based on Backbeat (replication, lifecycle, OOB updates, notification, bucket
policies, etc.) are not ready for 2.0 and must also be removed.

Took this as an opportunity to move Metadata Search into bucket operations, as
this is the context for MD searches.