ZEKO-3747 - sort first level keys of `deps.yaml`
In order to make it easier to find manually a component in this list
keep the list sorted (only the first level of keys).
This makes the build of the ISO sorted too, each image is pulled
in a sorted order.

in order to sort the keys use the following command:
`yq eval 'sortKeys(.)' deps.yaml`