ZENKO-3747 - add dashboard to images during build
When building the zenko ISO, add the OCI images that
contain the dashboards and alerts.

Images are pulled using `skopeo`, they are pushed on the registry
using `oras` during the component build step. oras produces a manifest
file with no `mediaType`, add it manually so the `static-container-registry`
can pickup the images (the `mediaType` is required).