Set wtimeout to 30s if not set
With no `wtimeout` option, when the level of write concern is
unachievable, the write operation will block indefinitely: which will
prevent APIs from returning, and possibly lead to a the frontend proxy
returning a timeout error to the client.

`wtimeout` solves this, as it causes write operations to return with an
error after the specified limit, even if the required write concern will
eventually succeed.

However, When these write operations return, MongoDB does not undo
successful data modifications performed before the write concern
exceeded the wtimeout time limit : which may put the database in a
unstable (data is written, but app does not know) or unsafe  state (data
is written to primary only, not yet replicated).

Issue: ZENKO-4606