ZENKO-4813: remove legacy identity logic
CTST used to embed the current identity, without support for
multiple ones at once, in the Cache (both as a cache key, or in
the cloned World's parameters). Eventually we end up with code
where the identity management is not clear: we have the cliMode,
the world.parameters, Cache.parameters and Cache.AccessKeys.
In the new version, CTST introduces a new class, Identity, that
manages all the identities of different types, and used in every
API calls dynamically. This class serves as a store and cache of
all identities created within a cucumberJS process, but also as
a way to switch between identities freely and more simply, while
abstracting the CLI logic  (the legacy "cliMode" option).