ZENKO-4813: tag kafkacleaner with dmf
This test needs some messages in kafka to properly execute. This
makes it dependent on other tests: hence, we must add their tag
as we updated the pickles to run only one type of scenario at a
time.

We use the @DMf tag only to run this test between both tests