Make sure the right Kafka images are used
Use a tree-hash in the registry, to ensure we pull the exact image that
was build, even in case multiple builds are running in parallel.

We use a tree hash instead of the Git hash to avoid pushing a new image
for every commit, as these do not often change.

Issue: ZENKO-4876