Deny PutObject action on DR site
This commit adds a step in the PRA feature test to deny the "PutObject" action on the DR site. It also includes a corresponding step definition in the common.ts file to assert that the access is denied. This change is necessary to ensure that the user is unable to perform the "PutObject" action on the DR site.

Issue: ZENKO-4871