Extend the duration checks for the DMF volume
In some cases, the platform will take some time to process the
event, so the test will fail, although there is no error.
This change also add a sleep between two checks, to reduce the
impact on the CPU when multiple tests are running this step in
parallel.

Issue: ZENKO-4806