Fix zkop install to support sha1
When an SHA1 is used which is not the tip of the branch, it is not
available after a `git clone` (which fetches only commits reachable
from current branches).

This causes issues if a build is started while the component branch with
the commit has been rewritten (rebase...)

To overcome this, the safe way is to explicitely fetch the commit: which
works for any commit (branch, tag, sha1).

Issue: ZENKO-4876