Update how zenko-ctst account is created
It seems that we are not robust against parallel account
creations in Vault/Pensieve, leading in some cases to
two accounts in the __account collection, but also in the
pensieve overlay.
However, when the overlay is updated (adding a new
location, or a bucket configuration), ZKOP will thus fail
the overlay validation and the reconciliations will all
fail. It was affecting the bucket tests, but may affect
more tests later.

Issue: ZENKO-4941