remove retry count from test ring bucket name
Configuration of the test environement is done through an image built in the build-test-image job,
this image contains the list of location details used to initialize the locations. This list isn't
updated when only retrying the end to end test jobs, and the bucket is create using the env var passed
to the pod. So we endup creating the bucket with the updated retry count but configure the location
with the old bucket name which fails as pensieve won't be able to find that bucket.

Issue: ZENKO-4414