Use ubuntu-22.04 runner for doc
Need to fix the package list (enchant --> enchant-2), and took the
change to cleanup:
- sort package list
- use package list from Dockerfile in gha (instead of duplicating it)
- update docs dockerfile to using ubuntu 22.04 as well
- fix tox.ini config

Upgrading to ubuntu 24.04 (with python 3.12) does not work easily: it
fails in sphinxcontrib.spelling module, which uses the 'imp' module
which has been removed in 3.12. The fix would be to bump all sphinx
packages.

Issue: ZENKO-4987