Do not change shard0 matchlabel for the PV
To avoid having to delete the STS for existing
deployments, we must avoid changing the matchSelectors
for the shard 0. Removing only the shard selector is
not enough, as we risk selecting volumes from other
shards.

Instead, we choose to modify the app name to include,
only for new shards, the new shard ID. Shard 0 won't
be updated, and the new shards will have their own
labels.

Issue: ZENKO-4641