Rework ReplicationLagWarning alert
The previous version was creating alerts after a fresh install:
- The oplog would not be full in this case
- No operation might cause the lag to be wrongly high, or
- ...the platform is in a transient / startup phase

We would create this alert even if the oplog was not full, because
the delta was too small. We here avoid that by ensuring this
alert is only created if the oplog circular buffer is fully
utilized. This makes more sense because in other cases, it means
we have margin.

Issue: ZENKO-5047